backend/requirements.txt
django==4.2.15
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
youtube-transcript-api==0.6.2
openai==1.14.0
pymongo==4.6.2
djangorestframework-simplejwt==5.3.1
yt-dlp==2024.3.10
drf-nested-routers==0.93.5
certifi==2024.2.2 